requirements.txt
deap
numpy
Pillow
